requirements.txt
Django>=2.1.7
psycopg2>=2.7
pycrypto